// 05-12-23 UATsrc/src/app/models/bike-insu.Model.ts
export class BikeInfoModel {
    constructor(_BikeBrandCode: number, _BikeModelCode: number, _BikeVariantCode: number) {
        this.BikeBrandCode = _BikeBrandCode;
        this.BikeModelCode = _BikeModelCode;
        this.BikeVariantCode = _BikeVariantCode;
    }
    BikeBrandCode: number;
    BikeBrandDesc!: string;
    BikeModelCode: number;
    BikeModelDesc!: string;
    BikeVariantCode: number;
    BikeVariantDesc!: string;
    RegistrationYearCode!: number;
    RegistrationYearDesc!: string;
    RegistrationDate!: string;
    CubicCapacity!: number;
    SeatingCapacity!: number;
    VehicleFuelDesc!: string;
    VehicleFuelCode!: number;

}
// ...
export interface IBikeBrands {
    BikeBrandCode: number;
    BikeBrandDesc: string;
    InsuranceCateCode: number;
}
export interface IBikeModels {
    BikeModelCode: number;
    BikeModelDesc: string;
}
export interface IBikeVariants {
    BikeVariantCode: number;
    BikeVariantDesc: string;
    CubicCapacity: string;
    SeatingCapacity: string;
    FuelType: string;
    FuelTypeCode: string;
}