// 05-12-23 UATsrc/src/app/models/common.Model.ts
export interface IRegistrationYear {
  RegistrationYearCode: number;
  RegistrationYearDesc: string;
}
// ...
export interface errorLog {

  UserCode?: any;
  ApplicationNo?: any;
  CompanyName?: any;
  ControllerName?: any;
  MethodName?: any;
  ErrorCode?: any;
  ErrorDesc?: any;

}